% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_DE_surface_antigen.R
\name{bulk_DE_surface_antigen}
\alias{bulk_DE_surface_antigen}
\title{bulk_DE_surface_antigen}
\usage{
bulk_DE_surface_antigen(
cancer.type = 'liver hepatocellular carcinoma',
normal.tissue = 'LIVER',
octad_counts_data_path = "~/Downloads/octad.counts.and.tpm.h5",
output_file_name = "liver_output"
)
}
\arguments{
\item{cancer.type}{A character string specifying the cancer name.It should be selected from total_cancer_count. Case-sensitive.}

\item{normal.tissue}{A character string specifying the corresponding normal tissue for the cancer.It should be selected from normal.tissue.Case-sensitive.}

\item{octad_counts_data_path}{path for octad.counts.and.tpm.h5}

\item{output_file_name}{A character string specifying the desired outfile file name}
}
\value{
dataframe with maximum 30 bispecific antibody target pairs

Barplot pdf showing frequency of top markers

Boxplot pdf showing expression pattern of top markers across cancer, normal tissue and normal vital organs

Volcanoplot pdf for each pair
}
\description{
This function enlists the bispecific antibody target pairs for given cancer.
The Pairtable.csv and related plots will be saved as pdf in working directory.
}
