% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_seurat_object.R
\name{download_seurat_object}
\alias{download_seurat_object}
\title{Download Seurat object based on object name}
\usage{
download_seurat_object(object_name = "HCC_GSE151530",save_dir = getwd())
}
\arguments{
\item{object_name}{A character string specifying the Seurat object name. Case-sensitive. Just use as given in 'SEURAT_OBJECT' of show_singlecell_data().}

\item{save_dir}{By default it will be saved in working directory.}
}
\value{
Seurat object will be downloaded in working directory and also loaded in environment.
}
\description{
Before running this function,please run.\cr
SeuratData=bsabsfinder::show_singlecell_data()\cr\cr
This function downloads the Seurat object corresponding to the given SEURAT_OBJECT from show_singlecell_data().
}
