% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_bispecific_expression.R
\name{sc_bispecific_expression}
\alias{sc_bispecific_expression}
\title{sc_bispecific_expression}
\usage{
sc_bispecific_expression(so = HCC_GSE151530 ,ident = "Celltype_MajorLineage",geneA = "GPC3",geneB="MUC13")
}
\arguments{
\item{so}{It should be loaded SuratObject in enviornment}

\item{ident}{Identity column of the seurat obect}

\item{geneA}{character string specifying first marker of a pair}

\item{geneB}{character string specifying second marker of a pair}
}
\value{
dataframe showing the percentage of cell expressing marker A , marker B and both marker A and marker B

PDF containing diplot of seuratobject, featureplot showing marker A and marker B and dotplot showing expression of marker A and marker B
}
\description{
This function calculates the \% of cells expressing given markers in seurat obect.
}
